+++ angel_serialize_generator/CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2.0.2
+* Generates an `XFields` class with the serialized names of
+all fields in a model class `X`.
+
 # 2.0.1
 * Ensured that `List` is only transformed if
 it generically references a `Model`.